#4 - Send data back to clients

+++ src/commonMain/kotlin/entities/CircleEntity.kt
@@ -6,8 +6,8 @@ import kotlin.math.sqrt
 
 interface CircleEntity {
     val id: Long
-    var position: Pair<Double, Double>
-    var size: Double
+    var position: Pair<Float, Float>
+    var size: Float
     var color: Color
 }
