+++ src/commonMain/kotlin/entities/Food.kt
@@ -4,7 +4,7 @@ import resources.Color
 
 data class Food(
     override val id: Long,
-    override var position: Pair<Double, Double>,
-    override var size: Double,
+    override var position: Pair<Float, Float>,
+    override var size: Float,
     override var color: Color
 ): CircleEntity
