+++ src/commonMain/kotlin/entities/Player.kt
@@ -4,8 +4,8 @@ import resources.Color
 
 data class Player(
     override val id: Long,
-    override var position: Pair<Double, Double>,
-    override var size: Double,
+    override var position: Pair<Float, Float>,
+    override var size: Float,
     override var color: Color,
     var name: String
 ): CircleEntity
