+++ src/jsMain/kotlin/resources/ColorExtensions.kt
@@ -2,4 +2,4 @@ package resources
 
 import kotlinx.css.rgb
 
-inline val Color.rgb get() = rgb(r, g, b)
+inline val Color.rgb get() = rgb(r.toInt(), g.toInt(), b.toInt())
